% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRoxygenDescription.R
\name{addRoxygenDescription}
\alias{addRoxygenDescription}
\title{Add Roxygen Description to R Function}
\usage{
addRoxygenDescription(
  Model = "gpt-4-0613",
  SelectedCode = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Model}{A character string specifying the GPT model to be used. Default is "gpt-4-0613".}

\item{SelectedCode}{A logical value indicating whether to use the selected code in RStudio. Default is TRUE.}

\item{verbose}{Logical flag to indicate whether to display the generated text. Default is TRUE.}
}
\value{
A message indicating completion if `SelectedCode` is TRUE, otherwise the Roxygen-annotated code is copied to the clipboard.
}
\description{
This function adds a Roxygen description to an R function using the GPT-4 model.
   It can either take the selected code from RStudio or read from the clipboard.
}
\details{
Add Roxygen Description to R Function
}
\examples{
\dontrun{
# Option 1
# Select some text in RStudio and then run the rstudio addins
# Option 2
# Copy the text into your clipboard then execute
addRoxygenDescription(Model = "gpt-4-0613", SelectedCode = FALSE)
}
}
\author{
Satoshi Kume
}
