% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatAI4pdf.R
\name{chatAI4pdf}
\alias{chatAI4pdf}
\title{chatAI4pdf}
\usage{
chatAI4pdf(pdf_file_path, nch = 2000, verbose = TRUE)
}
\arguments{
\item{pdf_file_path}{The path to the PDF file to be summarized. Must be a string.}

\item{nch}{The maximum number of characters for the summary. Default is 2000.}

\item{verbose}{Logical flag to print the summary. Default is TRUE.}
}
\value{
A string containing the summarized text.
}
\description{
Reads a PDF file and summarizes its content using a specified Large Language Model (LLM).
}
\details{
Summarize PDF Text Using a Large Language Model

This function reads a PDF file and summarizes its content using a specified Large Language Model (LLM).
}
\examples{
\dontrun{

#Read GPT-4 Technical Report
pdf_txt <- pdftools::pdf_text("https://cdn.openai.com/papers/gpt-4.pdf")

#Extract some text
pdf_txt <- pdf_txt[1:3]

#Execute
summary_text <- chatAI4pdf("1403.2805.pdf", Model, nch, Summary_block)
}
}
\author{
Satoshi Kume
}
