% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHEX.R
\name{createHEX4R}
\alias{createHEX4R}
\title{Create Hex Stickers for R Packages}
\usage{
createHEX4R(description, package_name, n = 3)
}
\arguments{
\item{description}{A string. This is the description of the R package for which the stickers will be generated.}

\item{package_name}{A string. This is the name of the R package for which the stickers will be generated.}

\item{n}{An integer. This is the number of stickers to generate. The default value is 3.}
}
\value{
A list of EBImage objects. These are the hexadecimal stickers generated for the R package.
}
\description{
Generates hexadecimal stickers for R packages using a provided description and package name. The stickers are generated by the OpenAI API.
}
\details{
Create Hex Stickers for R Packages

This function generates hexadecimal stickers, also known as "hex stickers", for R packages using the OpenAI API.
This is an experimental function.
}
\examples{
\dontrun{
# Define a description for the R package
description = "hatAI4R provides a seamless interface for integrating OpenAI and other APIs into R."
package_name = "R package"

# Generate hex stickers
res <- createHEX4R(description, package_name)

# Display the stickers
Display(res)
}
}
\author{
Satoshi Kume
}
