% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImagePrompt_v1.R
\name{createImagePrompt_v1}
\alias{createImagePrompt_v1}
\title{Create Image Prompt version 1}
\usage{
createImagePrompt_v1(content, Model = "gpt-4o-mini", len = 200)
}
\arguments{
\item{content}{A character string describing the image to be generated. If not provided, the function will throw a warning and stop.}

\item{Model}{A character string specifying the AI model to be used for text generation.}

\item{len}{Integer specifying the maximum length of the text input.}
}
\value{
A character string that serves as the prompt for generating an image.
}
\description{
This function creates a prompt for generating an image from text using an AI model.
 This is an experimental function.
}
\details{
Create Image Prompt version 1
}
\examples{
\dontrun{
createImagePrompt_v1(content = "A Japanese girl animation with blonde hair.")
}

}
\author{
Satoshi Kume
}
