% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRcode.R
\name{createRcode}
\alias{createRcode}
\title{Create R Code from Clipboard Content and Output into the R Console}
\usage{
createRcode(
  Summary_nch = 100,
  Model = "gpt-4-0613",
  SelectedCode = TRUE,
  verbose = TRUE,
  SlowTone = FALSE
)
}
\arguments{
\item{Summary_nch}{The maximum number of characters for the summary.}

\item{Model}{The model to be used for code generation, default is "gpt-4-0613".}

\item{SelectedCode}{A logical flag to indicate whether to read from RStudio's selected text. Default is TRUE.}

\item{verbose}{A logical value indicating whether to print the result to the console, default is TRUE.}

\item{SlowTone}{A logical value indicating whether to print the result slowly, default is FALSE.}
}
\value{
Prints the generated R code to the R console.
}
\description{
Reads text from the clipboard and generates R code based on the given input, printing the result to the R console.
}
\details{
Create R Code from Selected Text or Clipboard Content and Output into the R Console

This function reads text from your selected text or clipboard, interprets it as a prompt, and generates R code based on the given input.
The generated R code is then printed to the R console with optional slow printing.
This function can be executed from RStudio's Addins menu.
}
\section{RStudio Addins}{
 This function can be added to RStudio's Addins menu for easy access.
}

\examples{
\dontrun{

# Copy the origin of R code to your clipboard then execute from RStudio's Addins.

}
}
\author{
Satoshi Kume
}
