% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepel_rstudioapi_2EN.R
\name{deepel_rstudioapi_2EN}
\alias{deepel_rstudioapi_2EN}
\title{Translate Selected Text to English via DeepL API}
\usage{
deepel_rstudioapi_2EN()
}
\value{
A message indicating the completion of the translation. No value is returned.
}
\description{
Translate the selected text in RStudio into English using the DeepL API with auto-detection of the source language.
}
\details{
Translate Selected Text to English via DeepL API

This function uses the rstudioapi to execute English translation of the selected text using the DeepL API.
It will attempt to auto-detect the language of the text and translate it into English.
First, select the text and execute "DeepL Translation into English" from addins.
As a result, the selected part will be translated into English.
}
\examples{
\dontrun{

#Select the following text: "La selección está traducida al inglés."
#Then, execute "DeepL Translation into English" from addins.

}
}
\author{
Satoshi Kume
}
