% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeePL.R
\name{DeePL}
\alias{DeePL}
\title{DeepL API on R}
\usage{
DeePL(Sentence, EN2JA = TRUE, Auth_Key = "Your Authentication Key")
}
\arguments{
\item{Sentence}{an input string.}

\item{EN2JA}{a logical.}

\item{Auth_Key}{a string, Your Authentication Key}
}
\value{
character
}
\description{
DeepL API on R
}
\author{
Satoshi Kume
}
