% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designPackage.R
\name{designPackage}
\alias{designPackage}
\title{designPackage}
\usage{
designPackage(Model = "gpt-4-0613", verbose = TRUE, SlowTone = FALSE)
}
\arguments{
\item{Model}{The OpenAI GPT model to use for text generation. Default is "gpt-4-0613".}

\item{verbose}{Logical flag to indicate whether to display the generated text. Default is TRUE.}

\item{SlowTone}{Logical flag to indicate whether to print the text slowly. Default is FALSE.}
}
\value{
Prints the proposed design and architecture based on the verbosity and tone speed settings.
}
\description{
Assist in proposing the overall design and architecture of an R package.
}
\details{
Design Package for R

This function assists in proposing the overall design and architecture of an R package.
It uses the OpenAI GPT model for text generation to assist in the design process.
The function reads the input from the clipboard.
}
\examples{
\dontrun{
# Copy the text into your clipboard then execute
designPackage(Model = "gpt-4-0613", verbose = TRUE, SlowTone = FALSE)
}
}
\author{
Satoshi Kume
}
