% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discussion_flow_v1.R
\name{discussion_flow_v1}
\alias{discussion_flow_v1}
\title{discussion_flow_v1: Interactions and Flow Control Between LLM-based Bots (LLBs)}
\usage{
discussion_flow_v1(
  issue,
  Domain = "bioinformatics",
  Model = "gpt-4-0613",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  language = "English",
  Summary_nch = 50,
  verbose = TRUE,
  sayENorJA = TRUE
)
}
\arguments{
\item{issue}{The issue to be discussed. Example: "I want to solve linear programming and create a timetable."}

\item{Domain}{The domain of the discussion, default is "bioinformatics".}

\item{Model}{The model to be used, default is "gpt-4-0613".}

\item{api_key}{The API key for OpenAI, default is retrieved from the system environment variable "OPENAI_API_KEY".}

\item{language}{The language for the discussion, default is "English".}

\item{Summary_nch}{The number of characters for the summary, default is 50.}

\item{verbose}{Logical, whether to print verbose output, default is TRUE.}

\item{sayENorJA}{Logical, whether to say in English or Japanese, default is TRUE.}
}
\value{
A summary of the conversation between the bots.
}
\description{
Simulates interactions and flow control between three different roles of LLM-based bots (LLBs).
}
\details{
Interactions and Flow Control Between LLM-based Bots (LLBs)

This function simulates the relationships between three different roles of LLM bots (LLB A, B, C) to
reproduce more realistic dialogues and discussions.
The function assumes the following three roles:
A (Beginner): Generates questions and summaries based on the content of the discussion provided by the user.
B (Expert): Professionally answers questions from A.
C (Peer Reviewer): Reviews the dialogue between A and B and proposes improvements.
The three parties independently call the OpenAI API according to their roles, maintain their conversation history,
and execute the processes of questioning, answering, and peer reviewing.
The concept of the domain is to give a field of talk settings to each LLB.
It is recommended to use a model with accuracy higher than GPT-4.
English is recommended as the language, but verification will also be conducted in Japanese, the author's native language.
}
\examples{
\dontrun{
issue <-  "I want to solve linear programming and create a timetable."

#Run Discussion with the domain of bioinformatics
discussion_flow_v1(issue)
}
}
\author{
Satoshi Kume
}
