% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{Display}
\alias{Display}
\title{Display Images}
\usage{
Display(img, write = FALSE)
}
\arguments{
\item{img}{A list of images. Each element of the list is expected to be an image of class "Image".}

\item{write}{A logical, if TRUE, the function will also write each image to a file. Defaults to FALSE.}
}
\value{
No return value, this function is used for its side effect of displaying images or writing images to file.
}
\description{
This function displays images that are stored in a list. Each image is expected to be of the "Image" class. If write parameter is set to TRUE, it also saves each image as a PNG file.
}
\details{
Display Images
}
\examples{
\dontrun{
# Assuming "images" is a list of Image objects
Display(images)
# To also save images to file
Display(images, write = TRUE)
}
}
\author{
Satoshi Kume
}
