% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractKeywords.R
\name{extractKeywords}
\alias{extractKeywords}
\title{extractKeywords}
\usage{
extractKeywords(Model = "gpt-4o-mini", verbose = TRUE, SlowTone = FALSE)
}
\arguments{
\item{Model}{The OpenAI GPT model to use for text generation. Default is "gpt-4o-mini".}

\item{verbose}{Logical flag to indicate whether to print the result. Default is TRUE.}

\item{SlowTone}{Logical flag to indicate the speed of the output. Default is FALSE.}
}
\value{
Prints the extracted keywords based on verbosity and tone speed settings.
}
\description{
Extract keywords from input text and output them in a comma-separated format.
}
\details{
Extract Keywords from Text

This function extracts keywords from the input text. It uses the OpenAI GPT model
for text generation to assist in the extraction process. The function reads the input
from the clipboard and outputs the extracted keywords in a comma-separated format.
}
\examples{
\dontrun{
# Option 1
# Select some text in RStudio and then run the rstudio addins
# Option 2
# Copy the text into your clipboard then execute
extractKeywords(Model = "gpt-4o-mini", verbose = TRUE, SlowTone = FALSE)
}
}
\author{
Satoshi Kume
}
