% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateImage4R.R
\name{generateImage4R}
\alias{generateImage4R}
\title{Generate Images from Text}
\usage{
generateImage4R(
  content,
  n = 3,
  size = "256x256",
  response_format = "url",
  Output_image = TRUE,
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{content}{A character string describing the image to be generated.}

\item{n}{Integer specifying the number of images to generate. Must be between 1 and 10.}

\item{size}{A character string specifying the size of the generated images. Must be one of "256x256", "512x512", or "1024x1024".}

\item{response_format}{A character string specifying the format in which the generated images are returned. Must be one of "url" or "b64_json".}

\item{Output_image}{Logical indicating whether to return the images or their URLs. If TRUE, the images are returned.}

\item{api_key}{A character string specifying the OpenAI API key. If not provided, the function will attempt to read it from the system environment.}
}
\value{
A list of images or their URLs, depending on the value of Output_image.
}
\description{
This function generates images from text using the OpenAI API.
}
\details{
Generate Images from Text
}
\examples{
\dontrun{
generateImage4R(content = "Japanese gal girl", n = 3, size = "256x256",
                response_format = "url", Output_image = TRUE,
                api_key = Sys.getenv("OPENAI_API_KEY"))
}
}
\author{
Satoshi Kume
}
