% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_riken_pressrelease_urls.R
\name{get_riken_pressrelease_urls}
\alias{get_riken_pressrelease_urls}
\title{get_riken_pressrelease_urls}
\usage{
get_riken_pressrelease_urls(year)
}
\arguments{
\item{year}{The year for which press release URLs are to be retrieved.}
}
\value{
A character vector of URLs for the press releases of the specified year.
}
\description{
Retrieves the URLs of press releases from the RIKEN website for a given year.
}
\details{
Get URLs of RIKEN Press Releases

This function retrieves the URLs of press releases from the RIKEN website for a given year.
}
\examples{
\dontrun{
get_riken_pressrelease_urls(2023)
}
}
\author{
Satoshi Kume
}
