% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mac.R
\name{is_mac}
\alias{is_mac}
\title{Check Mac OS (Darwin)}
\usage{
is_mac()
}
\value{
Logical value indicating whether the system is Mac OS. Returns TRUE if Mac OS, FALSE otherwise.
}
\description{
Determines if the current system is running on Mac OS (Darwin).
}
\details{
Check if the System is Running on Mac OS

This function determines whether the current system is running on Mac OS.
It utilizes the system information provided by R to check the operating system name,
and returns TRUE if it is Mac OS (Darwin), and FALSE otherwise.
}
\examples{
\dontrun{
if (is_mac()) {
  print("This is a Mac OS system.")
} else {
  print("This is not a Mac OS system.")
}
}
}
\author{
Satoshi Kume
}
