% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsub.R
\name{ngsub}
\alias{ngsub}
\title{ngsub}
\usage{
ngsub(text)
}
\arguments{
\item{text}{The input text from which extra spaces and newline characters need to be removed.}
}
\value{
Returns the modified text as a character string.
}
\description{
Remove extra spaces and newline characters from text.
}
\details{
Remove Extra Spaces and Newline Characters

This function removes extra spaces and newline characters from the given text.
It replaces sequences of multiple spaces with a single space and removes newline characters followed by a space.
}
\examples{
\dontrun{
ngsub("This is  a text \n with  extra   spaces.")
}
}
\author{
Satoshi Kume
}
