% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLLMviaionet.R
\name{print.multiLLM_result}
\alias{print.multiLLM_result}
\title{Print method for multiLLM_result}
\usage{
\method{print}{multiLLM_result}(x, ...)
}
\arguments{
\item{x}{A multiLLM_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for multiLLM_result
}
