% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcodeImprovements.R
\name{RcodeImprovements}
\alias{RcodeImprovements}
\title{Suggest Improvements to the R Code on Your Clipboard}
\usage{
RcodeImprovements(
  Summary_nch = 100,
  Model = "gpt-4o-mini",
  verbose = TRUE,
  SlowTone = FALSE
)
}
\arguments{
\item{Summary_nch}{An integer specifying the maximum number of characters for the summary. Default is 100.}

\item{Model}{A character string specifying the GPT model to be used. Default is "gpt-4o-mini".}

\item{verbose}{A logical value indicating whether to print the result. Default is TRUE.}

\item{SlowTone}{A logical value indicating whether to print the result slowly. Default is FALSE.}
}
\value{
No return value; the function prints the suggestions for code improvement.
}
\description{
This function uses LLM to analyze the R code from the clipboard and suggests improvements.
   The function can also control the verbosity and speed of the output.
}
\details{
Suggest Improvements for R Code
}
\examples{
\dontrun{
#Copy your function to your clipboard
RcodeImprovements(Summary_nch = 100, Model = "gpt-4o-mini")
}
}
\author{
Satoshi Kume
}
