% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeQuotations.R
\name{removeQuotations}
\alias{removeQuotations}
\title{Remove All Types of Quotations from Text}
\usage{
removeQuotations(text)
}
\arguments{
\item{text}{A character string from which quotations will be removed.}
}
\value{
A character string with all types of quotations removed.
}
\description{
This function takes a text string as input and removes all occurrences
of single, double, and back quotations marks.
}
\examples{
\dontrun{
removeQuotations("\"XXX'`\"YYY'`") # Returns "XXXYYY"
}
}
\author{
Satoshi Kume
}
