% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riken_pressrelease_textEmbedding.R
\name{riken_pressrelease_textEmbedding}
\alias{riken_pressrelease_textEmbedding}
\title{riken_pressrelease_textEmbedding}
\usage{
riken_pressrelease_textEmbedding(url, api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{url}{A string. The URL of the RIKEN press-release from which to extract the text.}

\item{api_key}{A string. The API key for the text embedding service.}
}
\value{
A data frame of the text from the RIKEN press-release and its text embeddings.
}
\description{
This function extracts text from a given RIKEN press-release URL and performs text embedding.
}
\details{
Extract text and perform text embedding from RIKEN press-release
}
\examples{
\dontrun{
Sys.setenv(OPENAI_API_KEY = "Your API key")
riken_pressrelease_textEmbedding("https://www.riken.jp/pr/press/2020/20201218_1/", "your_api_key")
}
}
\author{
Satoshi Kume
}
