% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakInEN.R
\name{speakInEN}
\alias{speakInEN}
\title{Speak Selected Text in English}
\usage{
speakInEN(systemVoice = "Alex")
}
\arguments{
\item{systemVoice}{The voice to be used for speech. Default is "Alex".}
}
\value{
A message indicating the completion of the speech is returned.
}
\description{
This function reads aloud the selected text in English using the MacOS system's 'say' command.
   It uses the voice specified by the parameter `systemVoice`. The default voice is "Alex". It only works on MacOS systems.
}
\details{
Speak Selected Text in English on MacOS System
}
\examples{
\dontrun{
  # Select some text in RStudio and then run the rstudio addin
}
}
\author{
Satoshi Kume
}
