% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakInJA.R
\name{speakInJA}
\alias{speakInJA}
\title{Speak Selected Text in Japanese}
\usage{
speakInJA()
}
\value{
A message indicating the completion of the speech is returned.
}
\description{
This function reads aloud the selected text in Japanese using the MacOS system's 'say' command.
   The function specifically uses the 'Kyoko' voice for the speech. It only works on MacOS systems.
}
\details{
Speak Selected Text in Japanese on MacOS System
}
\examples{
\dontrun{
  # Select some text in RStudio and then run the rstudio addins
}
}
\author{
Satoshi Kume
}
