% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakInJA_v2.R
\name{speakInJA_v2}
\alias{speakInJA_v2}
\title{Speak Clipboard in Japanese}
\usage{
speakInJA_v2()
}
\value{
A message indicating the completion of the speech is returned.
}
\description{
This function reads aloud the clipboard content in Japanese using the MacOS system's 'say' command.
   The function specifically uses the 'Kyoko' voice for the speech. It only works on MacOS systems.
}
\details{
Speak Clipboard in Japanese on MacOS System
}
\examples{
\dontrun{
  # Copy some text into your clipboard in RStudio and then run the function
  speakInJA_v2()
}
}
\author{
Satoshi Kume
}
