% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportIdeaGeneration.R
\name{supportIdeaGeneration}
\alias{supportIdeaGeneration}
\title{supportIdeaGeneration: Support Idea Generation from Selected Text or Clipboard.}
\usage{
supportIdeaGeneration(
  Model = "gpt-4-0613",
  SelectedCode = TRUE,
  verbose = TRUE,
  SlowTone = FALSE
)
}
\arguments{
\item{Model}{The OpenAI GPT model to use for text generation. Default is "gpt-4-0613".}

\item{SelectedCode}{Logical flag to indicate whether to use the selected text in RStudio editor. Default is TRUE.}

\item{verbose}{Logical flag to indicate whether to display the generated text. Default is TRUE.}

\item{SlowTone}{Logical flag to indicate whether to print the text slowly. Default is FALSE.}
}
\value{
Prints the generated ideas or concepts based on the verbosity and tone speed settings.
}
\description{
Assist in generating ideas or concepts.
}
\details{
Support Idea Generation from Selected Text or Clipboard Input

This feature helps you generate ideas or concepts based on input from your selected text or clipboard.
It uses the OpenAI GPT model for text generation to assist in the idea generation process.
The function reads the input from the clipboard.
}
\examples{
\dontrun{
# Option 1
# Select some text in RStudio and then run the rstudio addins
# Option 2
# Copy the text into your clipboard then execute
supportIdeaGeneration()
}
}
\author{
Satoshi Kume
}
