% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSummaryAsBullet.R
\name{TextSummaryAsBullet}
\alias{TextSummaryAsBullet}
\title{Summarize Text into Bullet Points from Clipboard}
\usage{
TextSummaryAsBullet(
  text = clipr::read_clip(),
  BulletPoints = 4,
  Model = "gpt-4-0613",
  temperature = 1,
  verbose = FALSE,
  SpeakJA = FALSE
)
}
\arguments{
\item{text}{A character vector containing the text to be summarized. Default is text from the clipboard.}

\item{BulletPoints}{An integer specifying the number of bullet points for the summary. Default is 4.}

\item{Model}{A string specifying the machine learning model to use for text summarization. Default is "gpt-4-0613".}

\item{temperature}{A numeric value between 0 and 1 indicating the randomness of the text generation. Default is 1.}

\item{verbose}{A logical value indicating whether to print the summary. Default is FALSE.}

\item{SpeakJA}{A logical value indicating whether to speak the summary in Japanese. Default is FALSE.}
}
\value{
The summarized text in bullet points is returned and also copied to the clipboard.
}
\description{
This function takes a text input and summarizes it into a specified number of bullet points.
}
\details{
Summarize Text into Bullet Points
}
\examples{
\dontrun{
TextSummaryAsBullet(text = "This is a sample text.", BulletPoints = 3)
}
}
\author{
Satoshi Kume
}
