% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat4R.R
\name{chat4R}
\alias{chat4R}
\title{Chat4R: Interact with gpt-3.5-turbo-16k (default) using OpenAI API}
\usage{
chat4R(content, api_key, Model = "gpt-3.5-turbo-16k", temperature = 1)
}
\arguments{
\item{content}{A character string containing the user's input message.}

\item{api_key}{A character string containing the user's OpenAI API key.}

\item{Model}{A character string specifying the GPT model to use (default: "gpt-3.5-turbo").}

\item{temperature}{Numeric value controlling the randomness of the model's output (default: 1).}
}
\value{
A data frame containing the response from the chatGPT model.
}
\description{
This function uses the OpenAI API to interact with the GPT-3.5-turbo model (default)
and generates responses based on user input.
}
\details{
Chat4R Function
}
\examples{
\dontrun{
api_key <- "your_api_key_here"
response <- chat4R("What is the capital of France?", api_key)

#Check the result
response
}

}
\author{
Satoshi Kume
}
