% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkErrorDet_JP.R
\name{checkErrorDet_JP}
\alias{checkErrorDet_JP}
\title{Check Error Details in Japanese}
\usage{
checkErrorDet_JP(
  input = clipr::read_clip(),
  Summary_nch = 100,
  Model = "gpt-4-0314",
  verbose = TRUE,
  SlowTone = FALSE
)
}
\arguments{
\item{input}{A string containing the error message to be analyzed, read from the clipboard by default.}

\item{Summary_nch}{An integer specifying the maximum number of characters for the summary.}

\item{Model}{A string specifying the model to be used, default is "gpt-4-0314".}

\item{verbose}{A logical value to control the verbosity of the output, default is TRUE.}

\item{SlowTone}{A logical value to control the printing speed of the output, default is FALSE.}
}
\value{
The function prints the guidance on how to fix the error message.
}
\description{
A function to analyze and provide guidance on how to fix an error message copied from the R console.
}
\details{
Check Error Details in Japanese

This function provides a way to check error details in R. It takes an error message from the R console,
executes the function, and shows how to fix the error in Japanese.
}
\examples{
\dontrun{
  # Copy the error message that you want to fix.
  checkErrorDet_JP()
  checkErrorDet_JP(language = "Japanese")
}
}
\author{
Satoshi Kume
}
