% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions4R.R
\name{completions4R}
\alias{completions4R}
\title{completions4R}
\usage{
completions4R(
  prompt,
  api_key = Sys.getenv("OPENAI_API_KEY"),
  Model = "text-davinci-003",
  max_tokens = 50,
  temperature = 1,
  simple = TRUE
)
}
\arguments{
\item{prompt}{The initial text that the model responds to.}

\item{api_key}{The API key for OpenAI. Defaults to the value of the environment variable "OPENAI_API_KEY".}

\item{Model}{The ID of the model to use, such as "text-davinci-003".}

\item{max_tokens}{The maximum number of tokens to generate.}

\item{temperature}{Controls the randomness of the generated text. A value close to 0 produces more deterministic output, while a higher value (up to 2) produces more random output.}

\item{simple}{If TRUE, returns the generated text without newline characters. If FALSE, returns the full response from the API.}
}
\value{
The generated text or the full response from the API, depending on the value of `simple`.
}
\description{
Generate text using OpenAI's API
}
\details{
Generate text using OpenAI's API

This function sends a request to OpenAI's API to generate text based on the provided prompt and parameters.
}
\examples{
\dontrun{
  Sys.setenv(OPENAI_API_KEY = "Your API key")

  prompt <- "Translate the following English text to French: '{}'"

  completions4R(prompt, api_key)
}
}
\author{
Satoshi Kume
}
