% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertScientificLiterature.R
\name{convertScientificLiterature}
\alias{convertScientificLiterature}
\title{convertScientificLiterature}
\usage{
convertScientificLiterature(Model = "gpt-4-0613", SelectedCode = TRUE)
}
\arguments{
\item{Model}{The OpenAI GPT model to use for text generation. Default is "gpt-4-0613".}

\item{SelectedCode}{Logical flag to indicate whether to read the input from RStudio's active document. Default is TRUE.}
}
\value{
Inserts the converted text into the RStudio active document if SelectedCode is TRUE, otherwise writes to the clipboard.
}
\description{
Convert input text into scientific literature.
}
\details{
Convert to Scientific Literature

This function assists in converting the input text into scientific literature.
It uses the OpenAI GPT model for text generation to assist in the conversion process.
The function reads the input either from the RStudio active document or the clipboard.
}
\examples{
\dontrun{
convertScientificLiterature(SelectedCode = FALSE)
}
}
\author{
Satoshi Kume
}
