% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpecifications4R.R
\name{createSpecifications4R}
\alias{createSpecifications4R}
\title{Create Specifications for R Function}
\usage{
createSpecifications4R(Model = "gpt-4-0613", verbose = TRUE, SlowTone = FALSE)
}
\arguments{
\item{Model}{A character string specifying the GPT model to be used. Default is "gpt-4-0613".}

\item{verbose}{A logical value indicating whether to print the output. Default is TRUE.}

\item{SlowTone}{A logical value indicating whether to print the output slowly. Default is FALSE.}
}
\value{
The function prints the generated specifications to the console.
}
\description{
This function generates specifications for an R function from your clipboard.
   It takes in a text input, model name, verbosity, and tone speed to generate the specifications.
}
\details{
Create Specifications for R Function
}
\examples{
\dontrun{
createSpecifications4R(input = "Your R function specification")
}
}
\author{
Satoshi Kume
}
