% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{Display}
\alias{Display}
\title{Display Images and Optionally Write Them to Disk}
\usage{
Display(img, write = FALSE, Rds = FALSE)
}
\arguments{
\item{img}{A list of EBImage objects. These are the images to display and optionally write to disk.}

\item{write}{A boolean. Defaults to FALSE. If TRUE, the function writes the images to disk as PNG files.}
}
\value{
This function doesn't return anything. It displays images and optionally writes them to disk.
}
\description{
Displays a list of EBImage objects. If write is TRUE, it also writes them to the disk.
}
\details{
Display Images and Optionally Write Them to Disk

This function displays a list of EBImage objects, and optionally writes them to disk as PNG files.
}
\examples{
\dontrun{
# Load an image
img <- list(EBImage::readImage(system.file("images", "sample-color.png", package="EBImage")))

# Display the image
Display(img)

# Display and write the image to disk
Display(img, write = TRUE)
}
}
\author{
Satoshi Kume
}
