% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{Display}
\alias{Display}
\title{Display Images}
\usage{
Display(img)
}
\arguments{
\item{img}{A list of images, as returned by the generateImage4R function.}
}
\description{
This function displays images that are returned by the generateImage4R function.
}
\details{
Display Images
}
\examples{
\dontrun{
# Generate images
results <- generateImage4R(content = "Japanese gal girl")
# Display the generated images
Display(results)
}
}
\author{
Satoshi Kume
}
