% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editImage4R.R
\name{editImage4R}
\alias{editImage4R}
\title{Edit Images Using OpenAI API}
\usage{
editImage4R(
  image,
  mask = "",
  prompt,
  n = 3,
  size = "256x256",
  response_format = "url",
  Output_image = TRUE,
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{image}{A character string, the file path of the image to be edited. Must be a valid PNG file, less than 4MB, and square.}

\item{mask}{A character string, an additional image file path. If not provided, the transparency of the image will be used as the mask.}

\item{prompt}{A character string, a text description of the desired images. The maximum length is 1000 characters.}

\item{n}{An integer, the number of images to generate. Must be between 1 and 10. Defaults to 3.}

\item{size}{A character string, the size of the generated images. Must be one of "256x256", "512x512", "1024x1024". Defaults to "256x256".}

\item{response_format}{A character string, the format in which the generated images are returned. Must be one of "url" or "b64_json". Defaults to "url".}

\item{Output_image}{A logical flag, whether to output the images.}

\item{api_key}{A character string, the API key for OpenAI.}
}
\value{
If Output_image is TRUE, returns a list of images; if Output_image is FALSE, returns a character vector of URLs.
}
\description{
This function edits images using the OpenAI API. Given an original image and a prompt, it creates an edited or extended image.
}
\details{
Edit Images Using OpenAI API
}
\examples{
\dontrun{
# Edit images
results <- editImage4R(image = "path_to_image.png", prompt = "A blue sky", n = 3, size = "256x256", response_format = "url", Output_image = TRUE, api_key = Sys.getenv("OPENAI_API_KEY"))
# Display the edited images
Display(results)
}
}
\author{
Satoshi Kume
}
