% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevisedText.R
\name{revisedText}
\alias{revisedText}
\title{Revised Scientific Text}
\usage{
revisedText(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, whether to output verbose messages, default is TRUE.}
}
\value{
Revised text or relevant message.
}
\description{
This function prompts the user to input text, revision comments, and additional background information.
   It then revises the text according to the comments and outputs the revised text.
}
\details{
Revise Scientific Text
}
\examples{
\dontrun{
 revisedText()
}
}
\author{
Satoshi Kume
}
