% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportIdeaGeneration.R
\name{supportIdeaGeneration}
\alias{supportIdeaGeneration}
\title{supportIdeaGeneration: Support Idea Generation from Clipboard.}
\usage{
supportIdeaGeneration(Model = "gpt-4-0613", verbose = TRUE, SlowTone = FALSE)
}
\arguments{
\item{Model}{The OpenAI GPT model to use for text generation. Default is "gpt-4-0613".}

\item{verbose}{Logical flag to indicate whether to display the generated text. Default is TRUE.}

\item{SlowTone}{Logical flag to indicate whether to print the text slowly. Default is FALSE.}
}
\value{
Prints the generated ideas or concepts based on the verbosity and tone speed settings.
}
\description{
Assist in generating ideas or concepts.
}
\details{
Support Idea Generation from Clipboard.

This feature helps you generate ideas or concepts based on input from your clipboard.
It uses the OpenAI GPT model for text generation to assist in the idea generation process.
The function reads the input from the clipboard.
}
\examples{
\dontrun{
supportIdeaGeneration()
}
}
\author{
Satoshi Kume
}
