% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textEmbedding.R
\name{textEmbedding}
\alias{textEmbedding}
\title{Text Embedding from OpenAI API}
\usage{
textEmbedding(text, api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{text}{The input text to get the embedding for. This should be a character string.}

\item{api_key}{The API key for the OpenAI API. This should be a character string.}
}
\value{
A character string representing the text embedding from the OpenAI API.
}
\description{
This function calls the OpenAI API to get the text embedding of the input text.
}
\examples{
\dontrun{
　Sys.setenv(OPENAI_API_KEY = "Your API key")
  textEmbedding("Hello, world!")
}
}
\author{
Satoshi Kume
}
