% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSummary.R
\name{TextSummary}
\alias{TextSummary}
\title{Summarize Long Text}
\usage{
TextSummary(
  text = clipr::read_clip(),
  nch = 2000,
  Summary_block = 150,
  Model = "gpt-3.5-turbo",
  temperature = 1,
  verbose = TRUE,
  SpeakJA = FALSE
)
}
\arguments{
\item{text}{A character vector containing the text to be summarized.
If not provided, the function will attempt to read from the clipboard.}

\item{nch}{Integer specifying the number of characters at which to split the input text for processing.}

\item{Summary_block}{Integer specifying the target number of characters for the summarized output.}

\item{Model}{A character string specifying the AI model to be used for text summarization. Default is "gpt-3.5-turbo".
Execution with GPT-4 is recommended.}

\item{temperature}{Numeric specifying the randomness of the AI model's output.}

\item{verbose}{A logical flag to print the message. Default is TRUE.}

\item{SpeakJA}{A logical flag to enable Japanese voice output. Default is FALSE.}
}
\value{
The summarized text is placed into the clipboard and the function returns the result of \code{clipr::write_clip}.
}
\description{
This function summarizes a long text using AI models like GPT-4.
   It is recommended to execute this function with GPT-4. This is an experimental function.
}
\details{
Summarize Long Text
}
\examples{
\dontrun{
TextSummary(text = c("This is a long text to be summarized.",
                     "It spans multiple sentences and goes into much detail."),
            nch = 1000, Summary_block = 100,
            Model = "gpt-3.5-turbo", temperature = 1)
}
}
\author{
Satoshi Kume
}
